Your account has been tagged.
Today I logged into my Github account and noticed the message: “Your account has been tagged. Because of this, your profile is hidden from the public. If you believe this is an error, please contact support to verify the status of your account. " I have not violated any behavioral restrictions, and the only projects I publish in my account are for debank. Due to the ban, I was unable to provide my identification. Can anyone in support lift the ban? Thank you.